veegil/requirements.txt
aniso8601
appdirs
asgiref
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dataclasses
Django==4.0.1
django-cors-headers
django-decouple
django-environ
django-mongoengine
djangorestframework
djangorestframework-simplejwt
djongo==1.3.6
dnspython
drf-yasg
google-auth
graphene
graphene-django
graphql-core
graphql-relay
idna
inflection
Jinja2
kubernetes
MarkupSafe
mongoengine
mypy
mypy-extensions
djangorestframework-api-key
oauthlib
packaging
pillow
promise
pyasn1
pyasn1_modules
pycparser
pycryptodome
PyJWT
pymongo==3.12.3 
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
rsa
six
sqlparse
text-unidecode
tutor
tutor-openedx
typing_extensions
uritemplate
urllib3
websocket-client
dj_database_url
django-request-logging
django-requestlogging 
gunicorn
whitenoise
psycopg2
psycopg2-binary
